+++ src/Schemes/Ftp.php
@@ -33,6 +33,17 @@ class Ftp extends AbstractHierarchicalUri implements Uri
     /**
      * @inheritdoc
      */
+    protected function assertSupportedScheme()
+    {
+        $scheme = $this->getScheme();
+        if ('' !== $scheme) {
+            parent::assertSupportedScheme();
+        }
+    }
+
+    /**
+     * @inheritdoc
+     */
     protected function isValid()
     {
         return null === $this->fragment->getContent()